requirements.txt
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
stable-baselines3==1.6.1
mujoco==2.3.2
gym==0.21.0
einops
wandb
numpy
seaborn
pandas
tqdm
scipy
statsmodels
pingouin
opencv-python
ruamel.yaml
scikit-video
tensorboard